Pepper_and_Potato_disease_classification/api/requirements.txt
tensorflow==2.15.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.5.0
matplotlib
numpy